==> R/04_exploratory_analysis.Rmd <==
---
title: "Exploratory analysis and QAQC"
author : "Jeremy Wicquart"
date: "`r Sys.Date()`"
output: 
  html_document:
    fig_caption: yes
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(
	echo = FALSE,
	fig.align = "center",
	fig.height = 18,
	fig.width = 12,
	message = FALSE,
	warning = FALSE
)
```

# Defining functions and packages

```{r base}

# 1. Source functions ----

source("functions/graphical_par.R")
source("functions/theme_graph.R")
source("functions/pred_vbgf.R")

# 2. Required packages ----

library(tidyverse)
library(kableExtra)
library(formattable)
library(readxl)
library(plotly)
library(rfishbase)

# 3. Set theme_graph() as the default ggplot theme ----

theme_set(theme_graph())

# 4. Load data ----

data_complete <- read.csv("./../data/02_back-calculated-size-at-age_morat-et-al.csv")

```

# 1. Data exploration

## 1.1 Tables

### 1.1.1 Data with back-calc.

```{r}

data_complete %>% 
  filter(!is.na(Li_sp_m)) %>% 
  summarize_at(vars("ID", "Species", "Family"), n_distinct, na.rm = TRUE) %>% 
  bind_rows(data_complete %>%
              filter(!is.na(Li_sploc_m)) %>% 
              summarize_at(vars("ID", "Species", "Family"), n_distinct, na.rm = TRUE), .) %>% 
  bind_rows(data_complete %>%
              summarize_at(vars("ID", "Species", "Family"), n_distinct, na.rm = TRUE), .) %>% 
  mutate(Type = c("Entire dataset", "Back-calculated data by sp. and loc.", "Back-calculated data by sp."), .before = 1) %>% 
  kable(., 
        col.names = c("", "Individual", "Species", "Family"), 
        caption = "Table 1. Comparison of numbers of individuals, species and families for the entire dataset and for back-calculated data (by species, and by species and location).") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))

```

### 1.1.2 Individual by species

```{r}

data_complete %>%
  filter(!is.na(Li_sp_m)) %>% 
  group_by(Species) %>% 
  summarise(n = length(unique(ID)),
        min_lencap = round(min(Lcpt, na.rm = TRUE), 0),
        max_lencap = round(max(Lcpt, na.rm = TRUE), 0),
        max_age = max(Agei, na.rm = TRUE)) %>% 
  kable(., 
        col.names = c("Species", "n", "Min length (TL, mm)", "Max length (TL, mm)", "Age max (years)"),
        caption = "Table 2. Number of individuals, minimum and maximum total length (mm) and maximum age, by species. Only individuals with back-calculated data (by species) are shown.") %>%
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  column_spec(1, italic = T)

```

### 1.1.3 Individuals by location

```{r}

data_complete %>%
  filter(!is.na(Li_sp_m)) %>% 
  group_by(Species, Location) %>% 
  summarise(n = length(unique(ID))) %>% 
  pivot_wider(names_from = Location, values_from = n) %>% 
  mutate(Gambiers = cell_spec(Gambiers, "html", color = ifelse(is.na(Gambiers), "white", "#446CB3")),
         Hao = cell_spec(Hao, "html", color = ifelse(is.na(Hao), "white", "#446CB3")),
         Marquesas = cell_spec(Marquesas, "html", color = ifelse(is.na(Marquesas), "white", "#446CB3")),
         Moorea = cell_spec(Moorea, "html", color = ifelse(is.na(Moorea), "white", "#446CB3")),
         Manuae = cell_spec(Manuae, "html", color = ifelse(is.na(Manuae), "white", "#446CB3")),
         Tuamotu = cell_spec(Tuamotu, "html", color = ifelse(is.na(Tuamotu), "white", "#446CB3"))) %>% 
  kable(., 
        format = "html", escape = FALSE,
        caption = "Table 3. Number of individuals of each species by location. Only individuals with back-calculated data (by species) are shown.") %>%
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  column_spec(1, italic = T)

```

### 1.1.4 Number of NA

```{r}

# 1. Number of NA by variable

data_complete %>%
  summarise_all(~(sum(is.na(.)))) %>% 
  t(.) %>% 
  as.data.frame() %>% 
  kable(., col.names = c("Number of rows with NA"), caption = "Table 4. Number of rows with Non-Available (NA) data by variable") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))

```

## 1.2 Plot

### 1.2.1 Agei *vs* Li (sp.)

```{r fig.width=15, fig.height=25, fig.cap="Figure 1. Relation between length (TL, mm) and age, by species."}

data_complete %>% 
  filter(!is.na(Li_sp_m)) %>% # Remove individuals with non back-calculated data
  ggplot(., aes(x = Agei, y = Li_sp_m)) +
  geom_point(color = col_color_graph, fill = col_fill_graph, size = 1, shape = 21)+
  theme(strip.text.x = element_text(size = 8, face = "italic"),
        strip.background = element_rect(colour = "black", fill = col_facet, size = 1),
        plot.title = element_text(colour = col_color_graph),
        plot.subtitle = element_text(colour = "black"),
        strip.text.y = element_text(angle = 360, face = "italic")) +
  facet_wrap(~Species, scales = "free", ncol = 5) +
  labs(x = "Age (years)", y = "Length (mm)")

```

### 1.2.2 Agei *vs* Li (sp. and loc.)

```{r fig.width=15, fig.height=45, fig.cap="Figure 2. Relation between length (TL, mm) and age, by species and location."}

data_complete %>% 
  filter(!is.na(Li_sploc_m)) %>% # Remove individuals with non back-calculated data
  ggplot(., aes(x = Agei, y = Li_sploc_m)) +
  geom_point(color = col_color_graph, fill = col_fill_graph, size = 1, shape = 21)+
  theme(strip.text.x = element_text(size = 8),
        strip.background = element_rect(colour = "black", fill = col_facet, size = 1),
        plot.title = element_text(colour = col_color_graph),
        plot.subtitle = element_text(colour = "black"),
        strip.text.y = element_text(angle = 360, face = "italic")) +
  facet_grid(Species~Location, scales = "free") +
  labs(x = "Age (years)", y = "Length (mm)")

```

# 2. VBGF results

## 2.1 Growth parameters (sp.)

```{r}

# 1. Growth parameters (on back-calculated data by species) ----

read.csv("./../data/03_back-calculated_vbgf_predictions_sp.csv") %>% 
  mutate(Estimate = paste0(round(Estimate, 3), " (", round(Est.Error, 3), ")")) %>% 
  pivot_wider("Species", names_from = Parameter, values_from = Estimate) %>% 
  select(Species, linf, k, t0) %>% 
  arrange(Species) %>% 
  kable(., 
        col.names = c("Species", "Linf", "K", "t0"), 
        caption = "Table 5. Values of Von Bertalanffy parameters estimated on back-calculated data through Bayesian framework by species. Standard deviation are the values in parentheses. Linf is expressed in TL and cm.") %>% 
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  column_spec(1, italic = T)

```

## 2.2 Growth parameters (sp. loc.)

```{r}

# 1. Growth parameters (on back-calculated data by species and locations) ----

read.csv("./../data/03_back-calculated_vbgf_predictions_sploc.csv") %>% 
  mutate(Estimate = paste0(round(Estimate, 3), " (", round(Est.Error, 3), ")")) %>% 
  pivot_wider(c("Species", Location), names_from = Parameter, values_from = Estimate) %>% 
  select(Species, Location, linf, k, t0) %>% 
  arrange(Species) %>% 
  kable(., 
        col.names = c("Species", "Location", "Linf", "K", "t0"), 
        caption = "Table 6. Values of Von Bertalanffy parameters estimated on back-calculated data through Bayesian framework by species and location. Standard deviation are the values in parentheses. Linf is expressed in TL and cm.") %>% 
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  column_spec(1, italic = T)

```

## 2.3 Growth curves (sp.)

```{r fig.width=15, fig.height=25, fig.cap="Figure 3. Growth curves on back-calculated data by species."}

# 1. Growth curves ----

read.csv("./../data/03_back-calculated_vbgf_fitted_sp.csv") %>% 
  ggplot(data = .) +
    geom_ribbon(aes(x = age, ymin = ypred_lq, ymax = ypred_uq), alpha = 0.5, fill = col_fill_graph, alpha = 0.6) +
    geom_line(aes(x = age, y = ypred_m), color = col_color_graph) +
    geom_point(data = data_complete %>%
                 filter(Species %in% unique(read.csv("./../data/03_back-calculated_vbgf_fitted_sp.csv")$Species)) %>% 
                 mutate(Li_sp_m = Li_sp_m/10), 
               aes(x = Agei, y = Li_sp_m), color = col_color_graph, fill = col_fill_graph, size = 1, shape = 21) +
    facet_wrap(~Species, scales = "free", ncol = 5) +
    lims(y = c(0, NA)) +
    theme(strip.text.x = element_text(face = "italic")) +
    labs(x = "Age (years)", y = "Length (TL, cm)") +
    guides(colour = guide_legend(override.aes = list(alpha = 1, size = 0.5)))

```

## 2.4 Growth curves (sp. loc.)

```{r fig.width=15, fig.height=45, fig.cap="Figure 4. Growth curves on back-calculated data by species and location."}

# 1. Growth curves ----

read.csv("./../data/03_back-calculated_vbgf_fitted_sploc.csv") %>% 
  ggplot(data = .) +
    geom_ribbon(aes(x = age, ymin = ypred_lq, ymax = ypred_uq, fill = Location), alpha = 0.5, show.legend = FALSE) +
    geom_line(aes(x = age, y = ypred_m, color = Location), show.legend = FALSE) +
    geom_point(data = data_complete %>%
                 filter(Species %in% unique(read.csv("./../data/03_back-calculated_vbgf_fitted_sploc.csv")$Species)) %>% 
                 mutate(Li_sploc_m = Li_sploc_m/10), 
               aes(x = Agei, y = Li_sploc_m, color = Location), size = 1, show.legend = FALSE) +
    facet_grid(Species~Location, scales = "free") +
    lims(y = c(0, NA)) +
    labs(x = "Age (years)", y = "Length (TL, cm)") +
    theme(strip.text.x = element_text(size = 8),
          strip.background = element_rect(colour = "black", fill = col_facet, size = 1),
          plot.title = element_text(colour = col_color_graph),
          plot.subtitle = element_text(colour = "black"),
          strip.text.y = element_text(angle = 360, face = "italic"))

```

# 3. Raw *vs* Back-calculation

```{r fig.height=14, fig.width=8, include=FALSE}

# 1. List of species with enough individuals for one location ----

species_list <- c("Abudefduf sexfasciatus", "Acanthurus achilles", "Acanthurus pyroferus", 
                  "Acanthurus triostegus", "Balistapus undulatus", "Centropyge flavissima",
                  "Chaetodon ornatissimus", "Chromis iomelas", "Ctenochaetus marginatus",
                  "Dascyllus aruanus", "Epinephelus hexagonatus", "Myripristis berndti",
                  "Odonus niger", "Ostorhinchus angustatus", "Ostorhinchus apogonoides",
                  "Pristiapogon taeniopterus", "Sargocentron microstoma", "Stegastes nigricans")

# 1. Transform the raw data ----

# 1.1 Points --

data_raw_points <- read.csv("./../data/02_back-calculated-size-at-age_morat-et-al.csv") %>%
  filter(Species %in% species_list) %>% 
  select(Family, Genus, Species, ID, Agecpt, Lcpt, Location, Observer) %>% 
  unique() %>%
  dplyr::group_by(Species, Location) %>%
  dplyr::mutate(n = length(unique(ID))) %>%
  filter(n >= 10) %>% # filter with at least 10 replicates
  ungroup() %>% 
  dplyr::mutate(Lcpt = Lcpt/10) %>%  # Convert to cm
  mutate(Type = "Population-level data")

# 1.2 Curves -- 

data_raw_curve <- read.csv("./../data/03_raw_vbgf_fitted_sp.csv") %>% 
  filter(Species %in% species_list) %>% 
  mutate(Type = "Population-level data")

# 2. Transform the back-calculated data ----

# 2.1 Points --

data_backcalc_points <- read.csv("./../data/02_back-calculated-size-at-age_morat-et-al.csv") %>% 
  filter(Species %in% species_list) %>% 
  dplyr::mutate(Li_sp_m = Li_sp_m/10) %>%  # Convert to cm
  mutate(Type = "Back-calculated data")

# 2.2 Curves --

data_backcalc_curve <- read.csv("./../data/03_back-calculated_vbgf_fitted_sp.csv") %>% 
  filter(Species %in% species_list) %>% 
  mutate(Type = "Back-calculated data")

# 2. Make the plot ----

ggplot() +
  # Raw data
  geom_ribbon(data = data_raw_curve, aes(x = Agei, ymin = Q2.5, ymax = Q97.5, fill = Type), alpha = 0.5) +
  geom_line(data = data_raw_curve, aes(x = Agei, y = Estimate, color = Type), show.legend = FALSE) +
  geom_point(data = data_raw_points, aes(x = Agecpt, y = Lcpt, color = Type), size = 1) +
  # Back-calculated data
  geom_ribbon(data = data_backcalc_curve, aes(x = age, ymin = ypred_lq, ymax = ypred_uq, fill = Type), alpha = 0.5) +
  geom_line(data = data_backcalc_curve, aes(x = age, y = ypred_m, color = Type), show.legend = FALSE) +
  geom_point(data = data_backcalc_points, aes(x = Agei, y = Li_sp_m, color = Type), size = 1) +
  # Appearance
  facet_wrap(~Species, scales = "free", ncol = 3) +
  lims(y = c(0, NA)) +
  theme(strip.text.x = element_text(face = "italic"),
        legend.position = "top",
        legend.title = element_blank()) +
  labs(x = "Age (years)", y = "Length (TL, cm)") +
  lims(y = c(0, NA), x = c(0, NA)) +
  scale_color_manual(values = c(col_color_graph, "#d91e18")) +
  scale_fill_manual(values = c(col_fill_graph, "#e74c3c")) +
  guides(colour = guide_legend(override.aes = list(alpha = 1, size = 0.5)))

# 3. Remove useless objects ----

rm(data_raw_points, data_raw_curve, data_backcalc_points, data_backcalc_curve)

```

```{r fig.width=9, fig.height=12, fig.cap="Figure 5. Comparison of growth curves by species and location (one location only by growth curve) between back-calculated data and raw data (age estimated and length at capture)."}

# 1. Transform the raw data ----

species_list2 <- species_list[!species_list %in% c("Balistapus undulatus", "Centropyge flavissima", "Chaetodon ornatissimus")]

# 1.1 Points --

data_raw_points <- read.csv("./../data/02_back-calculated-size-at-age_morat-et-al.csv") %>%
  filter(Species %in% species_list2) %>% 
  select(Family, Genus, Species, ID, Agecpt, Lcpt, Location, Observer) %>% 
  unique() %>%
  dplyr::group_by(Species, Location) %>%
  dplyr::mutate(n = length(unique(ID))) %>%
  filter(n >= 10) %>% # filter with at least 10 replicates
  ungroup() %>% 
  dplyr::mutate(Lcpt = Lcpt/10) %>%  # Convert to cm
  mutate(Type = "Raw")

# 1.2 Curves -- 

data_raw_curve <- read.csv("./../data/03_raw_vbgf_fitted_sploc.csv") %>% 
  filter(Species %in% species_list2) %>% 
  mutate(Type = "Raw")

# 2. Transform the back-calculated data ----

# 2.1 Points --

data_backcalc_points <- read.csv("./../data/02_back-calculated-size-at-age_morat-et-al.csv") %>% 
  filter(Species %in% species_list2,
         !(Species == "Abudefduf sexfasciatus" & Location == "Moorea"),
         !(Species == "Acanthurus triostegus" & Location == "Marquesas"),
         !(Species == "Myripristis berndti" & Location == "Gambiers"),
         !(Species == "Myripristis berndti" & Location == "Marquesas")) %>% 
  dplyr::mutate(Li_sploc_m = Li_sploc_m/10) %>%  # Convert to cm
  mutate(Type = "Back-calculated")

# 2.2 Curves --

data_backcalc_curve <- read.csv("./../data/03_back-calculated_vbgf_fitted_sploc.csv") %>% 
  filter(Species %in% species_list2,
         !(Species == "Abudefduf sexfasciatus" & Location == "Moorea"),
         !(Species == "Acanthurus triostegus" & Location == "Marquesas"),
         !(Species == "Myripristis berndti" & Location == "Gambiers"),
         !(Species == "Myripristis berndti" & Location == "Marquesas")) %>% 
  mutate(Type = "Back-calculated")

# 2. Make the plot ----

ggplot() +
  # Raw data
  geom_ribbon(data = data_raw_curve, aes(x = Agei, ymin = Q2.5, ymax = Q97.5, fill = Type), alpha = 0.5) +
  geom_line(data = data_raw_curve, aes(x = Agei, y = Estimate, color = Type), show.legend = FALSE) +
  #geom_point(data = data_raw_points, aes(x = Agecpt, y = Lcpt, color = Type), size = 1) +
  # Back-calculated data
  geom_ribbon(data = data_backcalc_curve, aes(x = age, ymin = ypred_lq, ymax = ypred_uq, fill = Type), alpha = 0.5) +
  geom_line(data = data_backcalc_curve, aes(x = age, y = ypred_m, color = Type), show.legend = FALSE) +
  #geom_point(data = data_backcalc_points, aes(x = Agei, y = Li_sploc_m, color = Type), size = 1) +
  # Appearance
  facet_wrap(~Species, scales = "free", ncol = 3) +
  lims(y = c(0, NA)) +
  theme(strip.text.x = element_text(face = "italic"),
        legend.position = "top",
        legend.title = element_blank()) +  labs(x = "Age (years)", y = "Length (TL, cm)") +
  lims(y = c(0, NA), x = c(0, NA)) +
  scale_color_manual(values = c(col_color_graph, "#d91e18")) +
  scale_fill_manual(values = c(col_fill_graph, "#e74c3c")) +
  guides(colour = guide_legend(override.aes = list(alpha = 1, size = 0.5)))

# 3. Save the plot ----

ggsave("./../figs/figure-2_comparison-growth-curves-raw-vas-back-calculated.png", height = 15, width = 10) # PNG
ggsave("./../figs/figure-2_comparison-growth-curves-raw-vas-back-calculated.pdf", height = 15, width = 10, device = cairo_pdf) # PDF

# 4. Remove useless objects ----

rm(data_raw_points, data_raw_curve, data_backcalc_points, data_backcalc_curve)

```

# 4. Comparison with literature

```{r}

# 1. Import file synthesizing the VBGF parameters from literature ----

vbgf_literature <- read_excel("./../data/00_von-bertalanffy-literature.xlsx", sheet = 1) %>% 
  dplyr::mutate(Linf = ifelse(Size_unit == "mm", Linf/10, Linf),
                Size_max = ifelse(Size_unit == "mm", Size_max/10, Size_max)) %>% # Convert all length values to cm
  select(-Family, -Genus) # Remove Family and Genus to add those level through fishbase

# 2. Check validity of species names ----

vbgf_literature <- vbgf_literature %>% 
  rowwise() %>% 
  dplyr::mutate(Species2 = ifelse(is_empty(validate_names(Species)) == TRUE, 
                           NA, 
                           validate_names(Species))) %>% # Correct the names through fishbase
  ungroup() %>% 
  dplyr::mutate(Species = str_replace_all(Species, c("Acanthurus chirugus" = "Acanthurus chirurgus",
                                                     "Scarus psitticus" = "Scarus psittacus"))) %>% # Correct unfound names
  filter(Species %in% unique(data_complete$Species)) %>% 
  mutate(Reference = paste0(Reference, " (", Reference_type, ")")) %>%
  select(-Size_unit, -Sex, -Reference_type, -Species2) %>% 
  arrange(Species) %>% 
  filter(Species %in% unique(read.csv("./../data/03_back-calculated_vbgf_predictions_sp.csv")$Species))

```

```{r}

# 1. Get age max and length max by species ----

table_agemax <- data_complete %>% 
  drop_na(Li_sp_m) %>% 
  select(Species, Agecpt, Lcpt) %>% 
  group_by(Species) %>% 
  summarise(Age_max = max(Agecpt, na.rm = TRUE),
            Size_max = round(max(Lcpt, na.rm = TRUE)/10, 3))

# 2. Join mean estimated parameters, with size and length max, and with literature ----

read.csv("./../data/03_back-calculated_vbgf_predictions_sp.csv") %>% 
  mutate(Estimate = round(Estimate, 3)) %>% 
  pivot_wider(c("Species"), names_from = Parameter, values_from = Estimate) %>% 
  select(Species, linf, k, t0) %>% 
  rename(Linf = linf, K = k) %>% 
  mutate(Reference = "This study") %>% 
  left_join(., table_agemax) %>% 
  full_join(., vbgf_literature) %>%
  select(Species, Linf, K, t0, Age_max, Size_max, Reference) %>% 
  arrange(Species) %>% 
  write.csv2(., "./../figs/online-only-table-3_comparison-vbgf-parameters-literature.csv", row.names = FALSE)

```

```{r fig.height=15, fig.width=30, include=FALSE}

# 1. Comparison of VBGF parameters between our study and literature ----

vbgf_literature %>% 
  select(Species, Linf, K, t0, Reference) %>% 
  pivot_longer(c("Linf", "K", "t0"), names_to = "Parameter", values_to = "Estimate") %>% 
  mutate(Estimate = as.numeric(Estimate),
         Type = "Literature") %>% 
  bind_rows(., read.csv("./../data/03_back-calculated_vbgf_predictions_sp.csv") %>% 
                select(-Est.Error, -Q2.5, -Q97.5) %>% 
                mutate(Parameter = str_replace_all(Parameter, c("linf" = "Linf",
                                                                "k" = "K")),
                       Type = "This study",
                       Reference = "This study")) %>% 
  mutate(Species = as.factor(Species),
         Parameter = factor(Parameter, levels = c("Linf", "K", "t0"))) %>% 
  ggplot(data = ., aes(x = reorder(Species, desc(Species)), y = Estimate, color = Type)) +
    geom_point() +
    facet_wrap(~Parameter, scales = "free") +
    coord_flip() +
    labs(x = NULL, y = "Value") +
    theme(axis.text.y = element_text(face = "italic"))

```

```{r fig.height=10, fig.width=15, include=FALSE}

# 1. Get predictions from literature data ----

vbgf_literature_pred <- vbgf_literature %>% 
  # Complete missing Age_max
  filter(is.na(Age_max)) %>% 
  left_join(., vbgf_literature %>% 
              group_by(Species) %>% 
              summarise(Age_max2 = max(Age_max, na.rm = TRUE)) %>% 
              mutate(Age_max2 = na_if(Age_max2, -Inf))) %>% 
  mutate(Age_max = ifelse(is.na(Age_max), Age_max2, Age_max)) %>% 
  select(-Age_max2) %>% 
  bind_rows(., vbgf_literature %>% 
              filter(!(is.na(Age_max)))) %>% 
  arrange(Species) %>% 
  # Fix missing t0 to 0
  dplyr::mutate(t0 = ifelse(is.na(t0), 0, t0)) %>% 
  filter(!(is.na(K)), !(is.na(Linf)), !(is.na(t0))) %>% 
  mutate_at(c("Linf", "K", "t0"), as.numeric) %>% 
  filter(Species %in% unique(read.csv("./../data/03_back-calculated_vbgf_fitted_sp.csv")$Species)) %>% 
  # Make the predictions
  dplyr::mutate(Line = row_number()) %>% 
  group_by(Line) %>% 
  do(pred_vbgf(data = .))

# 2. Make the plot ----

plot_final <- read.csv("./../data/03_back-calculated_vbgf_fitted_sp.csv") %>% 
  filter(Species %in% unique(vbgf_literature_pred$Species)) %>% 
  ggplot(data = .) +
  geom_line(data = vbgf_literature_pred, aes(x = Agei, y = Li, group = as.factor(Line), 
                                             text = Reference), color = "#6c7a89", size = 0.7) +
  geom_ribbon(aes(x = age, ymin = ypred_lq, ymax = ypred_uq), alpha = 0.5, fill = col_fill_graph) +
  geom_line(aes(x = age, y = ypred_m), size = 0.7, color = col_color_graph) +

  facet_wrap(~Species, ncol = 4, scales = "free") +
  labs(x = "Age (years)", y = "Length (cm)") +
  theme(strip.text.x = element_text(face = "italic")) +
  lims(y = c(0, NA), x = c(0, NA)) +
  guides(colour = guide_legend(override.aes = list(alpha = 1, size = 0.5)))

plot_final

```

```{r fig.height=10, fig.width=9, fig.cap="Figure 6. Comparison of growth curves based on VBGF growth parameters estimated through Bayesian framework on species back-calculated data with growth curves from literature (interactive plot)."}

# 4. Convert to plotly ----

plot_final <- plot_final +
  theme(legend.position = "none")

ggplotly(plot_final, tooltip = c("Reference", "Location", "Size_type"))

```

# 5. Annex

```{r include=FALSE}

# 1. Variables description ----

data.frame(Column = 1:ncol(data_complete),
           Variable = colnames(data_complete),
           Description = NA,
           Unit = NA,
           Type = str_to_title(as.vector(sapply(data_complete, class)))) %>% 
  left_join(., data_complete %>% 
              mutate_if(is.character, ~NA) %>% 
              summarise_all(., ~paste0(round(min(., na.rm = TRUE), 3), " - ", round(max(., na.rm = TRUE), 3))) %>% 
              pivot_longer(everything(), names_to = "Variable", values_to = "Range") %>% 
              mutate(Range = na_if(Range, "Inf - -Inf"))) %>% 
  write.csv2(., "./../figs/table-2_variables-description.csv", row.names = FALSE)

```

```{r include=FALSE}

# 1. Extract the characteristics of individual used as otoliths section examples ----

sections_example <- data_complete %>% 
  filter(ID %in% c("GAM18_B006", "AC_LI_MA_03_17_221", "AC_NI_MA_03_17_118", "AC_PY_MA_03_17_009", "MOO18FE167",
                   "GAM18_B112", "CT_MA_MA_03_17_396", "GAM18_A033", "GAM18_A059", "GAM18_B080", # Acanthuridae
                   "OS_AN_MO_03_16_003", "OS_AP_MA_03_17_284", "PR_TA_MA_03_17_101", # Apogonidae
                   "OD_NI_MO_03_16_013", "BA_UN_MO_03_16_003", # Balistidae
                   "CA_ME_MO_03_16_005", # Carangidae
                   "GAM18_A100", "CH_CI_MO_03_16_012", # Chaetodontidae
                   "GAM18_A156", "SA_MI_MA_03_17_128", # Holocentridae
                   "GAM18_B066", "GAM18_A011", "GAM18_A214", # Labridae
                   "GAM18_B105", "GAM18_B093", # Lethrinidae
                   "LU_GI_MA_03_17_058", "GAM18_B099", "LU_FU_MO_03_16_007", # Lutjanidae
                   "GAM18_B129", "MU_FL_MO_03_16_013B", # Mullidae
                   "GAM18_A135", "CE_BI_MO_03_16_002", "GAM18_B056", "GAM18_A162", "DA_AR_MO_03_16_011",
                   "GAM18_A036", "GAM18_A191", "CH_JO_MO_03_16_013", "ST_AL_MO_03_16_006", # Pomacentridae
                   "GAM18_A120", "SC_PS_TA_03_16_019", # Scaridae
                   "GAM18_B012", # Scombridae
                   "GAM18_A073", "GAM18_A089", "EP_FA_MA_03_17_027", "H33", "GAM18_A154", "H70", "GAM18_B052", # Serranidae
                   "GAM18_B019", "SI_SP_MO_03_16_008")) %>% # Siganidae
  filter(Agei == 0) %>% 
  select(Family, Species, ID, Agecpt, Lcpt, Location) %>% 
  arrange(Family, Species) %>% 
  mutate(Legend = paste0(row_number(), ". ", Species, " (", Agecpt, " years, ", Lcpt, " mm TL)"))

# 2. Export the txt file needed for the legend of the figure on otoliths sections ----

write.table(paste(sections_example$Legend, collapse = ", "), 
            "./../figs/text_legend-fig-otolith-sections.txt", 
            col.names = FALSE, row.names = FALSE)

```

```{r}

# 1. Merge the tables on growth parameters by species and by species and location in one table ----

table_final <- read.csv("./../data/03_back-calculated_vbgf_predictions_sp.csv") %>% 
  mutate(Location = "All") %>% 
  bind_rows(., read.csv("./../data/03_back-calculated_vbgf_predictions_sploc.csv")) %>% 
  select(-Est.Error) %>% 
  pivot_wider(names_from = Parameter, names_glue = "{Parameter}_{.value}", values_from = c(Estimate, Q2.5, Q97.5)) %>% 
  arrange(Species, Location) %>% 
  mutate_if(is.numeric, ~round(.x, digits = 3)) %>% 
  mutate(linf_range = paste0(linf_Q2.5, " ; ", linf_Q97.5),
         k_range = paste0(k_Q2.5, " ; ", k_Q97.5),
         t0_range = paste0(t0_Q2.5, " ; ", t0_Q97.5)) %>% 
  select(Species, Location, linf_Estimate, linf_range, k_Estimate, k_range, t0_Estimate, t0_range) %>% 
  # Remove rows of only one sploc (-> sp sufficient) 
  group_by(Species) %>% 
  mutate(n = n()) %>% 
  filter(n > 2 | Location == "All") %>% 
  select(-n)

# 2. Create the table of number of individuals used to estimates parameters ----

# 2.1 Table for sp --

table_n_sp <- data_complete %>% 
  drop_na(Li_sp_m) %>%
  group_by(ID) %>%
  filter(Agecpt > 2) %>% # filter with maxage > 2
  ungroup() %>%
  dplyr::group_by(Species) %>%
  dplyr::summarise(n = length(unique(ID))) %>%
  filter(n >= 5) %>% # filter with at least 5 individuals
  ungroup() %>% 
  mutate(Location = "All")

# 2.1 Table for sploc --

table_n_sploc <- data_complete %>% 
  drop_na(Li_sploc_m) %>%
  group_by(ID) %>%
  filter(Agecpt > 2) %>% # filter with maxage > 2
  filter(!(ID %in% c("SA_MI_MA_03_17_128", "CE_FL_MA_03_17_214", "CE_FL_MA_03_17_216"))) %>% # Filter ID leading to non-convergence
  ungroup() %>%
  dplyr::group_by(Species, Location) %>%
  dplyr::summarise(n = length(unique(ID))) %>%
  filter(n >= 5) %>% # filter with at least 5 individuals
  ungroup()

# 2.3 Merge the n tables --

table_n <- full_join(table_n_sp, table_n_sploc) %>% 
  arrange(Species, Location)

# 3. Merge the n table with the 

table_final %>% 
  left_join(., table_n) %>% 
  select(Species, Location, n, linf_Estimate, linf_range, k_Estimate, k_range, t0_Estimate, t0_range) %>% 
  write.csv2(., 
            "./../figs/online-only-table-2_von-bertalanffy-parameters-estimates.csv", 
            row.names = FALSE)

```

# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Jeremy WICQUART | jeremywicquart@gmail.com | `r format(Sys.time())`
